# [Xamarin.Android.Build.Tasks] combine CheckTargetFrameworks/ResolveAssemblies