@jonpryor 2018-09-11
This saves 77ms.

We should mention the savings in the commit messages so I can do less math. :-)